camdaFisica.py
