--- graficos/apper_graph_manchester.py
@@ -1,31 +1,0 @@
-import gi
-import numpy as np
-
-gi.require_version("Gtk", "3.0")
-from gi.repository import Gtk
-
-import matplotlib.pyplot as plt
-from matplotlib.backends.backend_gtk3agg import FigureCanvasGTK3Agg as FigureCanvas
-from matplotlib.figure import Figure
-
-
-class Apper_graph_manchester:
-    def __init__(self, a = []):
-        self.a = a
-        self.apper_new_graph()
-
-    def apper_new_graph(self):
-        fig = Figure(figsize=(5, 4), dpi=100)
-        ax = fig.add_subplot(111)
-        t = np.arange(len(self.a))
-
-        # Plotando o sinal digital NRZ
-        ax.step(t, self.a, where='post', label='Sinal Manchester', linewidth=2)  # 'post' garante que o valor seja mantido após a transição
-        ax.set_title("Sinal Digital Manchester")
-        ax.set_xlabel("Tempo")
-        ax.set_ylabel("Valor")
-        ax.legend()
-        #Se quiser voltar para o codigo anterior tire as "" e
-        #essas duas linhas abaixo
-        canvas = FigureCanvas(fig)
-        return canvas
