--- list.py
@@ -1,5 +1,0 @@
-alist = []
-alist.append(1)
-alist.append(2)
-alist.append(3)
-print(alist)
